[Program] Fix null content

+++ TheGodfather/TheGodfather.cs
@@ -352,7 +352,7 @@ namespace TheGodfather
             }
 
             // Check if message contains filter
-            if (!e.Author.IsBot && e.Message.Content.Split(' ').Any(s => Commands.Messages.CommandsFilter.ContainsFilter(e.Guild.Id, s))) {
+            if (!e.Author.IsBot && e.Message.Content != null && e.Message.Content.Split(' ').Any(s => Commands.Messages.CommandsFilter.ContainsFilter(e.Guild.Id, s))) {
                 try {
                     await e.Channel.DeleteMessageAsync(e.Message);
                     _client.DebugLogger.LogMessage(
